Add files via upload
Hey! 

I created a Static HTML website for upcoming Durga Puja in Devi Pakhha. If you download the folder on your computer and click on the .html site of Durga Puja.html and increase the volume, you can enjoy the dhak music. Below every page I've included a hyperlink that redirects to the previous/next page of the website.

Hope you'll like it!